+++ Libplanet.Tests/Blockchain/Renderers/DelayedRendererTest.cs
@@ -19,7 +19,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
         protected static readonly IReadOnlyList<Block<DumbAction>> _chainA;
         protected static readonly IReadOnlyList<Block<DumbAction>> _chainB;
         protected static readonly Block<DumbAction> _branchpoint;
-        protected IComparer<BlockPerception> _canonicalChainComparer;
+        protected IComparer<IBlockExcerpt> _canonicalChainComparer;
         protected IStore _store;
         protected ILogger _logger;
 
@@ -76,7 +76,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
                 _chainB
             );
 
-            _canonicalChainComparer = new TotalDifficultyComparer(TimeSpan.FromSeconds(3));
+            _canonicalChainComparer = new TotalDifficultyComparer();
 
             _store = new DefaultStore(null);
             foreach (Block<DumbAction> b in _chainA.Concat(_chainB))